Compute shader almost finished!

+++ Engine/AtlasEngineGI/include/render/process/computeshaderprocess.h
@@ -14,10 +14,10 @@
  * */
 
 struct couleurStruct {
-    double r;
-    double g;
-    double b;
-    double a;
+    float r;
+	float g;
+	float b;
+	float a;
 };
 
 /*
